+++ app/src/main/java/sugar/free/sightremote/activities/boluses/ExtendedBolusActivity.java
@@ -143,7 +143,7 @@ public class ExtendedBolusActivity extends SightActivity implements TaskRunner.R
     }
 
     @Override
-    public void onAmountChange(BolusAmountPicker bolusAmountPicker, float newValue) {
+    public void onAmountChange(BolusAmountPicker bolusAmountPicker, double newValue) {
         deliver.setEnabled(newValue >= preperationResult.getMinBolusAmount() && durationPicker.getPickerValue() > 0);
     }
