+++ sightparser/src/main/java/sugar/free/sightparser/applayer/messages/remote_control/ExtendedBolusMessage.java
@@ -8,7 +8,7 @@ public class ExtendedBolusMessage extends BolusMessage {
     private static final long serialVersionUID = 1L;
 
     @Setter
-    private float amount;
+    private double amount;
     @Setter
     private int duration;
 
@@ -20,11 +20,11 @@ public class ExtendedBolusMessage extends BolusMessage {
         data.putShort((short) 0x1F00);
         data.putShort((short) 0x0000);
         data.putShort((short) 0x0000);
-        data.putUInt16LE((int) (amount * 100F));
+        data.putUInt16LE((int) (amount * 100D));
         data.putUInt16LE(duration);
         data.putShort((short) 0x0000);
         data.putShort((short) 0x0000);
-        data.putUInt16LE((int) (amount * 100F));
+        data.putUInt16LE((int) (amount * 100D));
         data.putUInt16LE(duration);
         return data.getBytes();
     }
