+++ sightparser/src/main/java/sugar/free/sightparser/applayer/messages/remote_control/MultiwaveBolusMessage.java
@@ -8,9 +8,9 @@ public class MultiwaveBolusMessage extends BolusMessage {
     private static final long serialVersionUID = 1L;
 
     @Setter
-    private float amount;
+    private double amount;
     @Setter
-    private float delayedAmount;
+    private double delayedAmount;
     @Setter
     private int duration;
 
@@ -21,12 +21,12 @@ public class MultiwaveBolusMessage extends BolusMessage {
         data.putShort((short) 0xFC00);
         data.putShort((short) 0x1F00);
         data.putShort((short) 0x0000);
-        data.putUInt16LE((int) (amount * 100F));
-        data.putUInt16LE((int) (delayedAmount * 100F));
+        data.putUInt16LE((int) (amount * 100D));
+        data.putUInt16LE((int) (delayedAmount * 100D));
         data.putUInt16LE(duration);
         data.putShort((short) 0x0000);
-        data.putUInt16LE((int) (amount * 100F));
-        data.putUInt16LE((int) (delayedAmount * 100F));
+        data.putUInt16LE((int) (amount * 100D));
+        data.putUInt16LE((int) (delayedAmount * 100D));
         data.putUInt16LE(duration);
         return data.getBytes();
     }
